Reorganised checkbox and radio button classes
