+++ lib/govuk_design_system_formbuilder.rb
@@ -6,19 +6,22 @@ require 'govuk_design_system_formbuilder/elements/hint'
 require 'govuk_design_system_formbuilder/elements/label'
 require 'govuk_design_system_formbuilder/elements/input'
 require 'govuk_design_system_formbuilder/elements/date'
-require 'govuk_design_system_formbuilder/elements/collection_radio_buttons'
-require 'govuk_design_system_formbuilder/elements/radios/collection_radio_button'
-require 'govuk_design_system_formbuilder/elements/radios/fieldset_radio_button'
-require 'govuk_design_system_formbuilder/elements/collection_check_boxes'
-require 'govuk_design_system_formbuilder/elements/check_boxes/collection_check_box'
-require 'govuk_design_system_formbuilder/elements/check_boxes/label'
-require 'govuk_design_system_formbuilder/elements/check_boxes/hint'
-require 'govuk_design_system_formbuilder/elements/error_message'
-require 'govuk_design_system_formbuilder/elements/error_summary'
 require 'govuk_design_system_formbuilder/elements/select'
 require 'govuk_design_system_formbuilder/elements/submit'
 require 'govuk_design_system_formbuilder/elements/text_area'
 require 'govuk_design_system_formbuilder/elements/file'
+
+require 'govuk_design_system_formbuilder/elements/radios/collection'
+require 'govuk_design_system_formbuilder/elements/radios/collection_radio_button'
+require 'govuk_design_system_formbuilder/elements/radios/fieldset_radio_button'
+
+require 'govuk_design_system_formbuilder/elements/check_boxes/collection'
+require 'govuk_design_system_formbuilder/elements/check_boxes/check_box'
+require 'govuk_design_system_formbuilder/elements/check_boxes/label'
+require 'govuk_design_system_formbuilder/elements/check_boxes/hint'
+
+require 'govuk_design_system_formbuilder/elements/error_message'
+require 'govuk_design_system_formbuilder/elements/error_summary'
 
 require 'govuk_design_system_formbuilder/containers/form_group'
 require 'govuk_design_system_formbuilder/containers/fieldset'
