+++ lib/govuk_design_system_formbuilder/elements/check_boxes/check_box.rb
